Parent / Child linkage
- Expanded on the wordpress_parent_id addition by adding
wordpress_parent & wordpress_children nodes
- Link to parent & children (if applicable) of all fields that are
pulled in (note this will be done even for fields where it makes no
sense - but the field will just be null / [])
- Using Gatsby API createSchemaCustiomization to link these nodes
(unlike all others linkages, which make use of the ___NODE pattern).
- This way seems better (& more robust*) - will look at migrating all
linkages to this pattern eventually. *No crashes when no categories have
children and the wordpress_children field is being queried, for instance